% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inter.R
\name{plot_inter}
\alias{plot_inter}
\title{Interaction Plot for Causal Forest}
\usage{
plot_inter(
  c.forest,
  x_var,
  y_var,
  bin_count = 50,
  limits = c(min(c.forest$predictions), max(c.forest$predictions))
)
}
\arguments{
\item{c.forest}{A fitted causal forest model object from the \code{grf} package.}

\item{x_var}{The name of the variable to be plotted on the x-axis.}

\item{y_var}{The name of the variable to be plotted on the y-axis.}

\item{bin_count}{Number of bins for the hexagonal binning. Default is 50.}

\item{limits}{The limits of the color scale, by default the minimum and maximum of the CATEs.}
}
\value{
A ggplot2 object with the interaction plot.
}
\description{
Generates an interaction plot for a causal forest model.
}
\examples{
\donttest{
library(grf)
set.seed(1995)
n <- 200; p <- 5
X <- matrix(rnorm(n * p), n, p)
colnames(X) <- paste0("X", seq_len(p))
W <- rbinom(n, 1, 0.5)
Y <- X[, 1] * W + rnorm(n)
cf <- causal_forest(X, Y, W, num.trees = 100)
plot_inter(cf, x_var = "X1", y_var = "X2")
}
}
