% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perm.R
\name{plot_perm}
\alias{plot_perm}
\title{Plot Variable Importance Scores from Permutation Analysis}
\usage{
plot_perm(perm_results)
}
\arguments{
\item{perm_results}{A list containing the output of the \code{left_perm} function, including the \code{plot_df} data frame.}
}
\value{
A ggplot2 object showing the distribution of variable importance scores.
}
\description{
Plots the distribution of variable importance scores from the \code{left_perm} function.
}
