% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_perm_methods.R
\name{print.cf_perm}
\alias{print.cf_perm}
\title{Print PermuCATE Variable Importance}
\usage{
\method{print}{cf_perm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cf_perm"} returned by \code{\link[=cf_perm]{cf_perm()}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The \code{cf_perm} object (invisibly).
}
\description{
Print PermuCATE Variable Importance
}
\seealso{
\code{\link[=cf_perm]{cf_perm()}}, \code{\link[=plot.cf_perm]{plot.cf_perm()}}
}
