% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.causal_forest.R
\name{print.summary.causal_forest}
\alias{print.summary.causal_forest}
\title{Print Summary of a Causal Forest}
\usage{
\method{print}{summary.causal_forest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.causal_forest"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x} (invisibly).
}
\description{
Print Summary of a Causal Forest
}
