% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rank.R
\name{rank_plot}
\alias{rank_plot}
\title{Rank Plot of Individual CATE Estimates}
\usage{
rank_plot(c.forest, show_ate_region = TRUE)
}
\arguments{
\item{c.forest}{A fitted causal forest model object from the \code{grf} package.}

\item{show_ate_region}{Logical indicating whether to show the ATE region (confidence interval). Default is TRUE.}
}
\value{
A ggplot2 object with the rank plot of individual CATE estimates.
}
\description{
Generates a rank plot of individual CATE (Conditional Average Treatment Effect)
estimates from a fitted causal forest model.
}
\examples{
\dontshow{if (rlang::is_installed(c("ggplot2", "ggdist"))) withAutoprint(\{ # examplesIf}
library(grf)
set.seed(1995)
n <- 200; p <- 5
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
Y <- X[, 1] * W + rnorm(n)
cf <- causal_forest(X, Y, W, num.trees = 100)
rank_plot(cf)
\dontshow{\}) # examplesIf}
}
