% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates_methods.R
\name{summary.cf_gates}
\alias{summary.cf_gates}
\title{Summarize GATES Results}
\usage{
\method{summary}{cf_gates}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cf_gates"} returned by \code{\link[=gates]{gates()}}.}

\item{...}{Additional arguments passed to \code{\link[=print.cf_gates]{print.cf_gates()}}.}
}
\value{
The \code{cf_gates} object (invisibly).
}
\description{
Summarize GATES Results
}
