% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_perm_methods.R
\name{summary.cf_perm}
\alias{summary.cf_perm}
\title{Summarize PermuCATE Variable Importance}
\usage{
\method{summary}{cf_perm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cf_perm"} returned by \code{\link[=cf_perm]{cf_perm()}}.}

\item{...}{Additional arguments passed to \code{\link[=print.cf_perm]{print.cf_perm()}}.}
}
\value{
The \code{cf_perm} object (invisibly).
}
\description{
Summarize PermuCATE Variable Importance
}
\seealso{
\code{\link[=cf_perm]{cf_perm()}}
}
