% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_loco.R
\name{cf_loco}
\alias{cf_loco}
\title{Modified LOCO Variable Importance for Causal Forests}
\usage{
cf_loco(
  c.forest,
  variable.groups = NULL,
  group.by.corr = FALSE,
  corr.threshold = 0.5,
  seed = 1234
)
}
\arguments{
\item{c.forest}{A fitted causal forest object from the grf package.}

\item{variable.groups}{A list of variable groups. Each element of the list should contain the variable names of a group.}

\item{group.by.corr}{A logical indicating whether to group variables by correlation. Default is FALSE.}

\item{corr.threshold}{A numeric value between 0 and 1 indicating the correlation threshold for grouping variables. Default is 0.5.}

\item{seed}{An integer seed for reproducibility.}
}
\value{
A numeric vector of variable importance scores normalized to sum to 1.
}
\description{
Computes variable importance for causal forests using the method here: https://arxiv.org/abs/2006.16763
}
