% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vimp.R
\name{compute_vimp}
\alias{compute_vimp}
\title{Compute Variable Importance}
\usage{
compute_vimp(alpha, Y.centered, W.centered, tau.hat)
}
\arguments{
\item{alpha}{Coefficients alpha of the retrained forest.}

\item{Y.centered}{Centered outcome.}

\item{W.centered}{Centered treatment assignment.}

\item{tau.hat}{Original out-of-bag predictions.}
}
\value{
A numeric value representing variable importance.
}
\description{
Computes variable importance from coefficients alpha of the retrained forest, centered outcome and treatment assignment, and original oob predictions.
}
