% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_loco_methods.R
\name{plot.cf_loco}
\alias{plot.cf_loco}
\title{Plot LOCO Variable Importance}
\usage{
\method{plot}{cf_loco}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cf_loco"} returned by \code{\link[=cf_loco]{cf_loco()}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{ggplot} object.
}
\description{
Draws a horizontal lollipop chart of LOCO variable importance scores.
}
