% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag.R
\name{plot_diag}
\alias{plot_diag}
\title{Model Diagnostics for Causal Forest}
\usage{
plot_diag(c.forest)
}
\arguments{
\item{c.forest}{A fitted causal forest model object from the \code{grf} package.}
}
\value{
A list containing the following elements:
\item{out_check}{A ggplot2 object showing the density plot comparing real Y and predicted Y.}
\item{bal_perm_plot}{A ggplot2 object showing the treatment propensity plot.}
\item{cate_distro_plot}{A ggplot2 object showing the density plot of the distribution of CATEs.}
\item{rmse_out}{The Root Mean Squared Error (RMSE) for the random forest model predictions.}
}
\description{
Generates a multi-panel diagnostics figure for a causal forest model.
}
\examples{
\dontrun{
cf_model <- causal_forest(X, Y, W)
diagnostics <- cf_model_diagnostics(cf_model)
}
}
