% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdp.R
\name{plot_pdp}
\alias{plot_pdp}
\title{Partial Dependence Plot for Causal Forest}
\usage{
plot_pdp(
  c.forest,
  x_var,
  curve_fitter = TRUE,
  method = "loess",
  show_ate_region = TRUE,
  x.limits = NULL,
  y.limits = NULL
)
}
\arguments{
\item{c.forest}{A fitted causal forest model object from the \code{grf} package.}

\item{x_var}{The name of the variable to be plotted on the x-axis.}

\item{curve_fitter}{Logical indicating whether to include a curve fitter. Default is TRUE.}

\item{method}{The smoothing method to be used by \code{geom_smooth} if \code{curve_fitter} is TRUE. Default is "loess".}

\item{show_ate_region}{Logical indicating whether to show the ATE region (confidence interval). Default is TRUE.}

\item{x.limits}{x axis limits specified as c() vector. Defaults to range of X.}

\item{y.limits}{y axis limits specified as c() vector. Defaults to range of Y.}
}
\value{
A ggplot2 object with the partial dependence plot.
}
\description{
Generates a partial dependence plot for a causal forest model.
}
\examples{
\dontrun{
cf_model <- causal_forest(X, Y, W)
pdp_plot(cf_model, x_var = "X1")
}
}
