% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_loco_methods.R
\name{print.cf_loco}
\alias{print.cf_loco}
\title{Print LOCO Variable Importance}
\usage{
\method{print}{cf_loco}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cf_loco"} returned by \code{\link[=cf_loco]{cf_loco()}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The \code{cf_loco} object (invisibly).
}
\description{
Print LOCO Variable Importance
}
